editSpineag

+++ api-v1-0/craftPlantOnRidge.php
@@ -9,22 +9,28 @@ if (isset($_POST['userId']) && !empty($_POST['userId'])) {
     $channelId = 1; // VK
 
     if ($app->checkSessionKey($_POST['userId'], $_POST['sessionKey'])) {
-        try {
-            $result = $mainDb->query('DELETE FROM user_plant_ridge WHERE id='.$_POST['dbId'].' AND user_id = '.$_POST['userId']);
-            if ($result) {
-                $json_data['message'] = '';
-                echo json_encode($json_data);
-            } else {
-                $json_data['id'] = 2;
-                $json_data['status'] = 's241';
-                throw new Exception("Bad request to DB!");
-            }
-        }
-        catch (Exception $e)
-        {
-            $json_data['status'] = 's046';
-            $json_data['message'] = $e->getMessage();
+        $m = md5($_POST['userId'].$_POST['dbId'].$app->md5Secret());
+        if ($m != $_POST['hash']) {
+            $json_data['id'] = 6;
+            $json_data['status'] = 's369';
+            $json_data['message'] = 'wrong hash';
             echo json_encode($json_data);
+        } else {
+            try {
+                $result = $mainDb->query('DELETE FROM user_plant_ridge WHERE id=' . $_POST['dbId'] . ' AND user_id = ' . $_POST['userId']);
+                if ($result) {
+                    $json_data['message'] = '';
+                    echo json_encode($json_data);
+                } else {
+                    $json_data['id'] = 2;
+                    $json_data['status'] = 's241';
+                    throw new Exception("Bad request to DB!");
+                }
+            } catch (Exception $e) {
+                $json_data['status'] = 's046';
+                $json_data['message'] = $e->getMessage();
+                echo json_encode($json_data);
+            }
         }
     } else {
         $json_data['id'] = 13;
