+++ api-v1-0/removeFromInventory.php
@@ -9,25 +9,28 @@ if (isset($_POST['userId']) && !empty($_POST['userId'])) {
     $channelId = 1; // VK
 
     if ($app->checkSessionKey($_POST['userId'], $_POST['sessionKey'])) {
-        try {
-            // $result = $mainDb->delete('user_building',
-            //     ['user_id' => $_POST['userId'], 'id' => $_POST['dbId']],
-            //     ['int', 'int']);
-            $result = $mainDb->query('DELETE FROM user_building WHERE user_id='.$_POST['userId'].' AND id = '.$_POST['dbId']);
-             if ($result) {
-                $json_data['message'] = '';
-                echo json_encode($json_data);
-            } else {
-                $json_data['id'] = 2;
-                $json_data['status'] = 's144';
-                $json_data['message'] = 'bad query';
-            }
-        }
-        catch (Exception $e)
-        {
-            $json_data['status'] = 's145';
-            $json_data['message'] = $e->getMessage();
+        $m = md5($_POST['userId'].$_POST['dbId'].$app->md5Secret());
+        if ($m != $_POST['hash']) {
+            $json_data['id'] = 6;
+            $json_data['status'] = 's381';
+            $json_data['message'] = 'wrong hash';
             echo json_encode($json_data);
+        } else {
+            try {
+                $result = $mainDb->query('DELETE FROM user_building WHERE user_id=' . $_POST['userId'] . ' AND id = ' . $_POST['dbId']);
+                if ($result) {
+                    $json_data['message'] = '';
+                    echo json_encode($json_data);
+                } else {
+                    $json_data['id'] = 2;
+                    $json_data['status'] = 's144';
+                    $json_data['message'] = 'bad query';
+                }
+            } catch (Exception $e) {
+                $json_data['status'] = 's145';
+                $json_data['message'] = $e->getMessage();
+                echo json_encode($json_data);
+            }
         }
     } else {
         $json_data['id'] = 13;
